PMM-8155 Move restore from backup service to backups logic
