+++ services/management/backup/backups_service.go
@@ -28,7 +28,6 @@ import (
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	backupv1beta1 "github.com/percona/pmm/api/managementpb/backup"
-	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"gopkg.in/reform.v1"
@@ -73,168 +72,15 @@ func (s *BackupsService) RestoreBackup(
 	ctx context.Context,
 	req *backupv1beta1.RestoreBackupRequest,
 ) (*backupv1beta1.RestoreBackupResponse, error) {
-	var params *prepareRestoreJobParams
-	var jobID, restoreID string
 
-	err := s.db.InTransactionContext(ctx, nil, func(tx *reform.TX) error {
-		var err error
-		params, err = s.prepareRestoreJob(tx.Querier, req.ServiceId, req.ArtifactId)
-		if err != nil {
-			return err
-		}
-
-		restore, err := models.CreateRestoreHistoryItem(tx.Querier, models.CreateRestoreHistoryItemParams{
-			ArtifactID: req.ArtifactId,
-			ServiceID:  req.ServiceId,
-			Status:     models.InProgressRestoreStatus,
-		})
-		if err != nil {
-			return err
-		}
-
-		restoreID = restore.ID
-
-		service, err := models.FindServiceByID(tx.Querier, req.ServiceId)
-		if err != nil {
-			return err
-		}
-
-		var jobType models.JobType
-		var jobResultData *models.JobResultData
-		switch service.ServiceType {
-		case models.MySQLServiceType:
-			jobType = models.MySQLRestoreBackupJob
-			jobResultData = &models.JobResultData{
-				MySQLRestoreBackup: &models.MySQLRestoreBackupJobResult{
-					RestoreID: restoreID,
-				}}
-		case models.MongoDBServiceType:
-			jobType = models.MongoDBRestoreBackupJob
-			jobResultData = &models.JobResultData{
-				MongoDBRestoreBackup: &models.MongoDBRestoreBackupJobResult{
-					RestoreID: restoreID,
-				}}
-		case models.PostgreSQLServiceType,
-			models.ProxySQLServiceType,
-			models.HAProxyServiceType,
-			models.ExternalServiceType:
-			return errors.Errorf("backup restore unimplemented for service type: %s", service.ServiceType)
-		default:
-			return errors.Errorf("unsupported service type: %s", service.ServiceType)
-		}
-
-		job, err := models.CreateJobResult(tx.Querier, params.AgentID, jobType, jobResultData)
-		if err != nil {
-			return err
-		}
-
-		jobID = job.ID
-
-		return err
-	})
+	id, err := s.backupsLogicService.RestoreBackup(ctx, req.ServiceId, req.ArtifactId)
 	if err != nil {
-		return nil, err
-	}
-
-	if err := s.startRestoreJob(jobID, req.ServiceId, params); err != nil {
 		return nil, err
 	}
 
 	return &backupv1beta1.RestoreBackupResponse{
-		RestoreId: restoreID,
+		RestoreId: id,
 	}, nil
-}
-
-type prepareRestoreJobParams struct {
-	AgentID      string
-	ArtifactName string
-	Location     *models.BackupLocation
-	ServiceType  models.ServiceType
-	DBConfig     *models.DBConfig
-}
-
-func (s *BackupsService) prepareRestoreJob(
-	q *reform.Querier,
-	serviceID string,
-	artifactID string,
-) (*prepareRestoreJobParams, error) {
-	service, err := models.FindServiceByID(q, serviceID)
-	if err != nil {
-		return nil, err
-	}
-
-	artifact, err := models.FindArtifactByID(q, artifactID)
-	if err != nil {
-		return nil, err
-	}
-
-	location, err := models.FindBackupLocationByID(q, artifact.LocationID)
-	if err != nil {
-		return nil, err
-	}
-
-	dbConfig, err := models.FindDBConfigForService(q, service.ServiceID)
-	if err != nil {
-		return nil, err
-	}
-
-	agents, err := models.FindPMMAgentsForService(q, serviceID)
-	if err != nil {
-		return nil, err
-	}
-	if len(agents) == 0 {
-		return nil, errors.Errorf("cannot find pmm agent for service %s", serviceID)
-	}
-
-	return &prepareRestoreJobParams{
-		AgentID:      agents[0].AgentID,
-		ArtifactName: artifact.Name,
-		Location:     location,
-		ServiceType:  service.ServiceType,
-		DBConfig:     dbConfig,
-	}, nil
-}
-
-func (s *BackupsService) startRestoreJob(jobID, serviceID string, params *prepareRestoreJobParams) error {
-	locationConfig := &models.BackupLocationConfig{
-		PMMServerConfig: params.Location.PMMServerConfig,
-		PMMClientConfig: params.Location.PMMClientConfig,
-		S3Config:        params.Location.S3Config,
-	}
-
-	switch params.ServiceType {
-	case models.MySQLServiceType:
-		if err := s.jobsService.StartMySQLRestoreBackupJob(
-			jobID,
-			params.AgentID,
-			serviceID,
-			0,
-			params.ArtifactName,
-			locationConfig,
-		); err != nil {
-			return err
-		}
-	case models.MongoDBServiceType:
-		if err := s.jobsService.StartMongoDBRestoreBackupJob(
-			jobID,
-			params.AgentID,
-			0,
-			params.ArtifactName,
-			params.DBConfig,
-			locationConfig,
-		); err != nil {
-			return err
-		}
-	case models.PostgreSQLServiceType,
-		models.ProxySQLServiceType,
-		models.HAProxyServiceType,
-		models.ExternalServiceType:
-		return status.Errorf(codes.Unimplemented, "unimplemented service: %s", params.ServiceType)
-	default:
-		return status.Errorf(codes.Unknown, "unknown service: %s", params.ServiceType)
-	}
-
-	return nil
 }
 
 // ScheduleBackup add new backup task to scheduler.
